refactor(language-service): language_service_adapter -> adapters
This rename is done because we know have a file system adapter over a
project as well as the compiler adapter.
