+++ packages/language-service/ivy/adapters.ts
@@ -6,6 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
+/** @fileoverview provides adapters for communicating the ng compiler */
+
 import {NgCompilerAdapter} from '@angular/compiler-cli/src/ngtsc/core/api';
 import {absoluteFrom, AbsoluteFsPath, FileStats, FileSystem, PathSegment, PathString} from '@angular/compiler-cli/src/ngtsc/file_system';
 import {AdapterResourceLoader} from '@angular/compiler-cli/src/ngtsc/resource';
