+++ packages/language-service/ivy/compiler_factory.ts
@@ -12,7 +12,7 @@ import {TrackedIncrementalBuildStrategy} from '@angular/compiler-cli/src/ngtsc/i
 import {TypeCheckingProgramStrategy} from '@angular/compiler-cli/src/ngtsc/typecheck/api';
 import * as ts from 'typescript/lib/tsserverlibrary';
 
-import {LanguageServiceAdapter} from './language_service_adapter';
+import {LanguageServiceAdapter} from './adapters';
 import {isExternalTemplate} from './utils';
 
 export class CompilerFactory {
